Résolution de problèmes d'affichage de la validation de la commande et amélioration du responsive.

+++ js/ordered.js
@@ -1,39 +1,68 @@
 //Fichier javascript exclusivement réservé à la page de la validation de la commande du site Oriteddies
 
-//Variable qui récupère l'url de la page ordered.html avec le résultat comme paramètre
-let url = new URL(window.location.href);
-
 //Variables récupérant le prénom et le nom de l'utilisateur
-let prenom = url.searchParams.get("prenom");
-let nom = url.searchParams.get("nom");
+let prenom = localStorage.getItem("prenom");
+let nom = localStorage.getItem("nom");
 
 //Variable récupérant le prix total de la commande
-let total = url.searchParams.get("total");
+let total = localStorage.getItem("totalPrix");
 
 //Variable récupérant l'identifiant de la commande
-let id = url.searchParams.get("id");
+let id = localStorage.getItem("id");
 
 //Variable récupérant la boîte html pour afficher le résultat de la commande
 let validationCommande = document.getElementById("validationCommande");
 
-//Titre remerciant l'utilisateur pour sa commande
-let titreMerci = document.createElement("h2");
-titreMerci.setAttribute("id", "titreMerci");
-titreMerci.innerHTML = "Merci " + prenom + " " + nom + " pour votre commande.";
-validationCommande.appendChild(titreMerci);
+//Vérifier si les données saisies ne sont pas nulles
+
+let donneesValides ;
+
+if(nom == null){
+    donneesValides = false;
+}else if(prenom == null){
+    donneesValides = false;
+}else if(id == null){
+    donneesValides = false;
+}else if(total == null){
+    donneesValides = false;
+}else{
+    donneesValides = true;
+}
+
+if (localStorage.length == 0 || !(donneesValides)) {
+    let aucuneCommande = document.createElement("h2");
+    aucuneCommande.setAttribute("id", "aucuneCommande");
+    aucuneCommande.innerHTML = "Aucune commande n'a été passée.";
+    validationCommande.appendChild(aucuneCommande);
+} else {
+    validationCommande.appendChild(remerciement());
+    validationCommande.appendChild(afficherMontantCommande());
+    validationCommande.appendChild(afficherNumeroCommande());
+    //Suppression des données contenues dans le localStorage
+    //Le panier est vide
+    localStorage.clear();
+}
+
+//Paragraphe de remerciement pour la commande passé par l'utilisateur
+function remerciement(){
+    let remerciement = document.createElement("p");
+    remerciement.setAttribute("id", "remerciement");
+    remerciement.innerHTML = "Merci " + prenom + " " + nom + " pour votre commande.";
+    return remerciement;
+}
 
 //Paragraphe indiquant le montant de la commande
-let montantCommande = document.createElement("p");
-montantCommande.setAttribute("id", "montantCommande");
-montantCommande.innerHTML = "Montant de la commande : " + total + "€";
-validationCommande.appendChild(montantCommande);
+function afficherMontantCommande(){
+    let montantCommande = document.createElement("p");
+    montantCommande.setAttribute("id", "montantCommande");
+    montantCommande.innerHTML = "Montant de la commande : " + total + "€";
+    return montantCommande;
+}
 
 //Paragraphe indiquant le numéro de la commande
-let numeroCommande = document.createElement("p");
-numeroCommande.setAttribute("id", "numeroCommande");
-numeroCommande.innerHTML = "Commande n°" + id;
-validationCommande.appendChild(numeroCommande);
-
-//Suppression des données contenues dans le localStorage
-//Le panier est vide
-localStorage.clear();
+function afficherNumeroCommande(){
+    let numeroCommande = document.createElement("p");
+    numeroCommande.setAttribute("id", "numeroCommande");
+    numeroCommande.innerHTML = "Commande n°" + id;
+    return numeroCommande;
+}
